+++ lib/utils.dart
@@ -1,3 +1,2 @@
-export 'src/utils/config.dart';
 export 'src/utils/token.dart';
 export 'src/utils/utils.dart';